requirements.txt
streamlit
tensorflow
numpy
pandas
pillow
fpdf
matplotlib